patch done

+++ client-cms/src/assets/Home.jsx
@@ -4,9 +4,11 @@ import BASE_URL from "../constant";
 import Swal from "sweetalert2";
 import { redirect, useNavigate } from "react-router-dom";
 import ModalFormAddEdit from "./ModalFormAddEdit";
+import PatchDataImage from "./ModalPatchImage";
 
 function Home() {
   const navigate = useNavigate();
+  const [loading, setLoading] = useState(false);
   const [edit, setEdit] = useState(0);
   const [input, setInput] = useState({
     name: "",
@@ -18,24 +20,24 @@ function Home() {
   function clearEdit() {
     setEdit(0);
   }
-  const [deletecuisineparams, setDeletecuisine] = useState(0)
-  async function deleteCuisine(){
+  const [deletecuisineparams, setDeletecuisine] = useState(0);
+  async function deleteCuisine() {
     try {
-        const { data } = await axios({
-            method: "delete",
-            url: `${BASE_URL}cuisine/${deletecuisineparams}`,
-            headers: {
-              Authorization: `Bearer ` + localStorage.accessToken,
-            },
-        });
-        FetchData()
+      const { data } = await axios({
+        method: "delete",
+        url: `${BASE_URL}cuisine/${deletecuisineparams}`,
+        headers: {
+          Authorization: `Bearer ` + localStorage.accessToken,
+        },
+      });
+      FetchData();
     } catch (error) {
-        console.log(error);
+      console.log(error);
     }
   }
-  useEffect(()=>{
-    deleteCuisine()
-  }, [deletecuisineparams])
+  useEffect(() => {
+    if (deletecuisineparams) deleteCuisine();
+  }, [deletecuisineparams]);
   function handleChange(event) {
     const { name, value } = event.target;
     const newInput = {
@@ -46,20 +48,20 @@ function Home() {
   }
   const [add, setAdd] = useState("");
   const handleSubmit = async (event) => {
-    event.preventDefault()
+    event.preventDefault();
     try {
-        let config = {
-            method: "post",
-            url: "https://bismillah.watersnj.com/cuisine",
-            headers: {
-              Authorization: `Bearer ` + localStorage.accessToken,
-            },
-            data: input,
-        }
-        if(edit !== 0) {
-            config.method = "put",
-            config.url = "https://bismillah.watersnj.com/cuisine/" + edit
-        }
+      let config = {
+        method: "post",
+        url: "https://bismillah.watersnj.com/cuisine",
+        headers: {
+          Authorization: `Bearer ` + localStorage.accessToken,
+        },
+        data: input,
+      };
+      if (edit !== 0) {
+        (config.method = "put"),
+          (config.url = "https://bismillah.watersnj.com/cuisine/" + edit);
+      }
       const { data } = await axios(config);
       setAdd("done");
       setInput({
@@ -147,6 +149,41 @@ function Home() {
   useEffect(() => {
     FetchData();
   }, [add]);
+  const [currentImage, setCurrentImage] = useState(null);
+  const [idPatch, setIdPatch] = useState(0);
+  const HandleImageChange = (event) => {
+    const image = event.target.files[0];
+    setCurrentImage(image);
+  };
+  const HandlePatchImage = async (event) => {
+    event.preventDefault();
+    setLoading(true);
+    try {
+      const formData = new FormData();
+      formData.append("photo", currentImage);
+      const { data } = await axios({
+        method: "patch",
+        url: `${BASE_URL}cuisine/${idPatch}`,
+        headers: {
+          Authorization: `Bearer ` + localStorage.accessToken,
+          "Content-Type": "multipart/form-data",
+        },
+        data: formData,
+      });
+    } catch (error) {
+      console.log(error);
+    } finally {
+      setLoading(false);
+      FetchData();
+    }
+  };
+  if (loading) {
+    return (
+        <div className="spinner-border text-info container d-flex justify-content-center align-items-center" role="status">
+          <span className="sr-only">Loading...</span>
+        </div>
+    );
+  }
   return (
     <>
       <div className="container">
@@ -224,13 +261,39 @@ function Home() {
                     >
                       Edit
                     </button>
-                    <button 
-                    className="btn btn-danger mt-3 ms-3"
-                    onClick={() => {
-                        setDeletecuisine(item.id)
-                    }}
-                    >Delete</button>
-                    <button className="btn btn-primary mt-3 ms-2">Change Image</button>
+                    <button
+                      className="btn btn-danger mt-3 ms-3"
+                      onClick={() => {
+                        setDeletecuisine(item.id);
+                      }}
+                    >
+                      Delete
+                    </button>
+                    <button
+                      type="button"
+                      className="btn btn-primary mt-3 ms-2"
+                      data-bs-toggle="modal"
+                      data-bs-target="#Patch"
+                      onClick={() => {
+                        setIdPatch(item.id);
+                      }}
+                    >
+                      Change Image
+                    </button>
+                    <div
+                      className="modal fade"
+                      id="Patch"
+                      data-bs-backdrop="static"
+                      data-bs-keyboard="false"
+                      tabIndex={-1}
+                      aria-labelledby="staticBackdropLabel"
+                      aria-hidden="true"
+                    >
+                      <PatchDataImage
+                        HandleImageChange={HandleImageChange}
+                        HandlePatchImage={HandlePatchImage}
+                      />
+                    </div>
                   </td>
                 </tr>
               );
